build: upgrade to Modular 26.5.0 (Mojo 1.0.0 stable)
Migrates the whole codebase off Mojo 1.0.0b2/Modular 26.4.0. Covers several
real language changes (MutAnyOrigin banned in struct fields, relocated
std.gpu.host/std.algorithm imports, InlineArray no longer ImplicitlyCopyable,
DevicePassable now rejects bare Int/UInt kernel-launch args, GPU codegen now
rejects SIMD[DType.float32, 3] entirely), plus a narrow, whole-program-scale-
sensitive compiler bug that made sppm.mojo's GPU kernels fail to typecheck
(worked around by moving sppm_render_gpu and its kernels into bdpt.mojo,
where the same code compiles fine).

Also fixes two real bugs surfaced along the way: a use-after-free in
test_bdpt.mojo's BVH fixture, and a genuine InlineArray-copy bug in
restir_sms.mojo's sms_shift that had apparently never been exercised
end-to-end since the 1.0.0 migration began.

Verified: make release (0 errors) and make unittest (38/39 files clean) both
pass; real CPU, --gpu, --sppm --gpu, and --vcm (BDPT) renders all produce
correct output on real hardware.

+++ Tests/unit/test_gpu.mojo
@@ -2,7 +2,7 @@ from std.math import abs
 from std.sys.info import size_of
 from std.testing import assert_true, TestSuite
 from std.sys import has_accelerator
-from std.gpu.host import DeviceContext
+from max.gpu.host import DeviceContext
 from gonzales.gpu import clear_film_gpu, accumulate_film_gpu
 from gonzales.geometry import RGB, Point3f, Vec3f, Ray_C, PathState_C
 from gonzales.spectrum import SampledWavelengths
@@ -46,7 +46,7 @@ def test_clear_film_gpu_zeroes_a_dirty_buffer() raises:
     ctx.synchronize()
 
     ctx.enqueue_function[clear_film_gpu](
-        buf.unsafe_ptr(), n_pixels,
+        buf.unsafe_ptr(), Int64(n_pixels),
         grid_dim=1, block_dim=n_pixels,
     )
     ctx.synchronize()
@@ -86,7 +86,7 @@ def test_accumulate_film_gpu_adds_path_estimates_onto_existing_film() raises:
 
     ctx.enqueue_function[accumulate_film_gpu](
         path_buf.unsafe_ptr().bitcast[PathState_C](),
-        film_buf.unsafe_ptr(), albedo_buf.unsafe_ptr(), n,
+        film_buf.unsafe_ptr(), albedo_buf.unsafe_ptr(), Int64(n),
         grid_dim=1, block_dim=n,
     )
     ctx.synchronize()
